requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.2
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.16.1
environ==1.0
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
